Refactor error rendering
